requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
argcomplete=1.12.3=pyhd3eb1b0_0
async_generator=1.10=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.2=py37h2a96729_1
brotli=1.0.9=h8ffe710_6
brotli-bin=1.0.9=h8ffe710_6
ca-certificates=2021.10.8=h5b45459_0
certifi=2021.10.8=py37h03978a9_1
colorama=0.4.4=pyhd3eb1b0_0
console_shortcut=0.1.1=4
cycler=0.11.0=pyhd8ed1ab_0
debugpy=1.5.1=py37hd77b12b_0
decorator=5.1.0=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py37_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=h546665d_1
greenlet=1.1.2=py37hf2a7229_1
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
importlib-metadata=4.8.2=py37haa95532_0
importlib_metadata=4.8.2=hd3eb1b0_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.4.1=py37haa95532_1
ipython=7.29.0=py37hd4e2768_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jbig=2.1=h8d14728_2003
jedi=0.18.0=py37haa95532_1
jinja2=3.0.2=pyhd3eb1b0_0
jpeg=9e=h8ffe710_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter_client=7.1.0=pyhd3eb1b0_0
jupyter_core=4.9.1=py37haa95532_0
jupyterlab_pygments=0.1.2=py_0
kiwisolver=1.3.1=py37hd77b12b_0
lcms2=2.12=h2a16943_0
lerc=3.0=h0e60522_0
libbrotlicommon=1.0.9=h8ffe710_6
libbrotlidec=1.0.9=h8ffe710_6
libbrotlienc=1.0.9=h8ffe710_6
libclang=11.1.0=default_h5c34c98_1
libdeflate=1.8=h8ffe710_0
libpng=1.6.37=h1d00b33_2
libprotobuf=3.6.0=h1a1b453_0
libtiff=4.3.0=hd413186_2
libwebp=1.2.0=h2bbff1b_0
libzlib=1.2.11=h8ffe710_1013
lz4-c=1.9.3=h8ffe710_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.0.1=py37h2bbff1b_0
matplotlib=3.5.0=py37haa95532_0
matplotlib-base=3.5.0=py37h6214cd6_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py37hfa6e2cd_1001
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py37h2bbff1b_0
mkl_fft=1.3.1=py37h277e83a_0
mkl_random=1.2.2=py37hf11a4ad_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=pyh9f0ad1d_0
mysql-connector-c=6.1.11=h33f27b4_1
mysql-connector-python=8.0.18=py37h13ed8b8_1
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=5.5.0=py_0
nbformat=5.1.3=pyhd3eb1b0_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
notebook=5.7.10=py37_0
numexpr=2.8.1=py37hb80d3ca_0
numpy=1.21.2=py37hfca59bb_0
numpy-base=1.21.2=py37h0829f74_0
olefile=0.46=pyh9f0ad1d_1
openjpeg=2.4.0=hb211442_1
openssl=1.1.1l=h8ffe710_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py37h6214cd6_0
pandoc=2.12=haa95532_0
pandocfilters=1.4.3=py37haa95532_1
parso=0.8.3=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=py37hd45dc43_0
pip=21.2.4=py37haa95532_0
prometheus_client=0.12.0=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.6.0=py37he025d50_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.10.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py37h6538335_2
pyrsistent=0.18.0=py37h196d8e1_0
python=3.7.11=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.7=2_cp37m
pytz=2020.1=py_0
pywin32=302=py37h827c3e9_1
pywinpty=0.5.7=py37_0
pyzmq=22.3.0=py37hd77b12b_2
qt=5.9.7=vc14h73c81de_0
scipy=1.7.3=py37h0a974cb_0
seaborn=0.11.0=py_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=58.0.4=py37haa95532_0
sip=4.19.8=py37h6538335_0
six=1.16.0=pyhd3eb1b0_0
sqlalchemy=1.4.31=py37hcc03f2d_0
sqlite=3.37.0=h2bbff1b_0
terminado=0.9.4=py37haa95532_0
testpath=0.5.0=pyhd3eb1b0_0
tk=8.6.11=h8ffe710_1
tornado=6.1=py37h2bbff1b_0
traitlets=5.1.1=pyhd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py37_1
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py37haa95532_2
winpty=0.4.3=4
xz=5.2.5=h62dcd97_1
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.11=h8ffe710_1013
zstd=1.5.2=h6255e5f_0
